# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "College Hierarchy API",
    "version": "1.0.0",
    "description": "API documentation for the College Hierarchy system"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Server"
    }
  ],
  "paths": {
...
    "/hod/gethodbyid/{id}": {
      "get": {
        "summary": "Get HOD by ID",
        "description": "Retrieves a HOD by ID.",
        "operationId": "getHodById",
        "tags": ["HOD"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "HOD fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "HOD with id: 1 fetched successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "Dr. Pavan" },
                        "principalId": { "type": "integer", "example": 1 }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Invalid id format (controller)" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "HOD not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "HOD with 1 not found (controller)" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Error in getting HOD" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/hod/getallbyid/{id}": {
      "get": {
        "summary": "Get HOD and Child Data by ID",
        "description": "Retrieves a HOD along with associated Professors and Students by HOD ID.",
        "operationId": "getAllHodById",
        "tags": ["HOD"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "HOD and child data fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "HOD and child data fetched successfully" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Id is required" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "HOD not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "HOD not found or fetching failed" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Error while fetching HOD" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/hod/deletehod/{id}": {
      "delete": {
        "summary": "Delete HOD",
        "description": "Deletes a HOD and related child data.",
        "operationId": "deleteHod",
        "tags": ["HOD"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "HOD and related data deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "HOD and related child data deleted successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Id is required" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "HOD not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Id not found or delete failed" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Error while deleting HOD" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}